fix(rpa):loci makefile check-python pass
