+++ core/plugin/rpa/main.py
@@ -43,7 +43,6 @@ def load_env_file(env_file: str) -> None:
 
     print(f"📋 Loading configuration file: {env_file}")
 
-    use_polaris = False
     os.environ["CONFIG_ENV_PATH"] = env_file
     with open(env_file, "r", encoding="utf-8") as f:
         for line_num, line in enumerate(f, 1):
@@ -62,57 +61,8 @@ def load_env_file(env_file: str) -> None:
                 else:
                     print(f"CFG  ✅ {key.strip()}={value.strip()}")
 
-                if key.strip() == "USE_POLARIS" and value.strip() == "true":
-                    use_polaris = True
             else:
                 print(f"  ⚠️  Line {line_num} format error: {line}")
-
-    if not use_polaris:
-        return
-
-    print(f"🔧 Config: USE_POLARIS :{use_polaris}")
-    load_polaris()
-
-
-def load_polaris() -> None:
-    """
-    Load remote configuration and override environment variables
-    """
-    from common.settings.polaris import ConfigFilter, Polaris
-
-    base_url = os.getenv("POLARIS_URL")
-    project_name = os.getenv("PROJECT_NAME", "hy-spark-agent-builder")
-    cluster_group = os.getenv("POLARIS_CLUSTER", "")
-    service_name = os.getenv("SERVICE_NAME", "rpa-server")
-    version = os.getenv("VERSION", "1.0.0")
-    config_file = os.getenv("CONFIG_FILE", "config.env")
-    config_filter = ConfigFilter(
-        project_name=project_name,
-        cluster_group=cluster_group,
-        service_name=service_name,
-        version=version,
-        config_file=config_file,
-    )
-    username = os.getenv("POLARIS_USERNAME")
-    password = os.getenv("POLARIS_PASSWORD")
-
-    # Ensure required parameters are not None
-    if not base_url or not username or not password or not cluster_group:
-        return  # Skip polaris config if required params are missing
-
-    polaris = Polaris(base_url=base_url, username=username, password=password)
-    try:
-        _ = polaris.pull(
-            config_filter=config_filter,
-            retry_count=3,
-            retry_interval=5,
-            set_env=True,
-        )
-    except (ConnectionError, TimeoutError, ValueError) as e:
-        print(
-            f"⚠️ Polaris configuration loading failed, "
-            f"continuing with local configuration: {e}"
-        )
 
 
 def start_service() -> None:
